Added full search on /pokemon
Now there is no need to specify a category, the bot will perform a full search in all the available categories. I left the possibility to choose the category to avoid homonym like "Psychic" which is both a move and a type (even though there is no support for types output in the bot). Also made the category lowercase and replaced è and é with regular e.