Auto-translate for all googletrans languages
Added support for all of the available languages in googletrans.
Improved CmdTranslate.translate_message().
Fixed CmdTranslate.function() not using specified source language.
Note: parse_message() in handlers.py checks if auto-tr is not bool. This is needed for backward compatibility (chats' settings created before this update for which auto-tr is either True or False). Users who had it turned on need to update their preferences manually with /settings.