Fixed error when requesting all items of a category
On some categories (such as abilities), when asking pokeapi for 9999 items (?limit=9999) it returns "Forbidden". Now the bot performs two requests: the first one gets the exact number of items in the requested category (count), the second one uses that number to perform a clean request (?limit={count}). Also, open() functions were moved closer to the json.dump() function, in order to avoid the creation of empty files whenever an exception is raised before the writing of the content.